stress-fractal: use faster atomic inc on larger row counter
The spin lock on the row counter is a relaitvely heavyweight operation
when scaling to multiple threads. Use a faster atomic inc on a larger
64 bit row counter to replace lock/unlock with atomic increment. There
is a slight issue with the row counter wrapping around after 2^64
increments as this can result in rows being missed if the fractal
height does not divide into 2^64 exactly. However, the wrap around
will take ~15 billion years using the smallest 64 x 64 fractal, so
I think this is acceptable.

+++ stress-ng.h
@@ -637,7 +637,7 @@ typedef struct {
 	} atomic ALIGN64;		/* Shared atomic temp vars */
 	struct {
 		void *lock;
-		int32_t row;
+		int64_t row;
 	} fractal;
 	struct {
 		/* futexes must be aligned to avoid -EINVAL */
